==> scripts/strat/4_analisis_exploratorio-geom.Rmd <==
---
title: "Análisis Exploratorio"
author: "Juan Baeza Ruiz-Henestrosa"
date: "`r Sys.Date()`"
output:
  html_document:
    df_print: paged
  pdf_document: default
editor_options:
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
knitr::opts_knit$set(root.dir = "..") # Para que haga knitr desde el directorio del proyecto

```

```{r}
library(tidyverse)
library(skimr)
library(sf)
library(corrplot)
library(GGally) # Coordenadas paralelas
library(ggpubr)
library(mapSpain)
```


Cargamos el archivos de datos depurados y obtenido a partir de un "muestreo estratificado" por mes :

```{r}
load("salidas_intermedias/datos_strat_depurados_geom_2024-05-03.RData")
```


## Resumen numérico

```{r}
datos %>%  st_drop_geometry %>% skim(.data_name = "datos")
```

Se tienen variables numéricas, categóricas, así como fechas y cadenas de caracteres. 

Las variables `cod_municipio` y `municipio` son "identificadoras", por lo que no se utilizarán en los modelo que se construyan.
La variable `date` contiene la fecha de cada registro y la variable `geometry` las coordenadas del punto.


## Variable objetivo:

En primer lugar se estudiará gráficamente la distribución temporal y espacial de la variable objetivo.

### Distribución temporal
#### Mes
```{r}
g_mes <- datos %>% 
  st_drop_geometry() %>% 
  # filter(fire==1) %>% 
  count(month(date,label=T),fire) %>% 
  rename("mes" = "month(date, label = T)") %>% 
  ggplot(aes(x = mes,y = n,fill=fire)) +
  geom_col(position="dodge",alpha=0.8) +
  # theme(axis.text.x = element_text(angle = 90)) +
  scale_fill_hue(direction = -1) +
  theme_minimal() +
  xlab("Mes") +
  ylab("Número de observaciones") +
  theme(legend.position = "top")


g_mes
```

Se observa que las observaciones se concentran entorno a los meses de verano y que en cada mes hay una cantidad similar de observaciones positivas y negativas. Esto se debe al procedimiento de muestreo que se ha utilizado para generar la muestra, ya que se ha estratificado por mes, de forma que la proporción de observaciones en cada mes fuese igual en ambas clases. Esto se ha hecho así con el fin de evitar introducir sesgos en la muestra para que los modelos entrenados usándola permitan efectivamente ser útiles en la predicción de incendios forestales.



#### Año
```{r}
g_year <- datos %>% 
  st_drop_geometry() %>% 
  # filter(fire==1) %>% 
  count(year(date),fire) %>% 
  rename("año" = "year(date)") %>% 
  ggplot(aes(x = año,y = n,fill=fire)) +
  geom_col(position="dodge",alpha=0.8) +
  scale_x_continuous(breaks=2002:2022) +
  # geom_smooth(se=F,aes(color=fire),alpha=0.1) +
  # scale_color_manual(values=c("darkblue","darkred")) + 
  scale_fill_hue(direction = -1) +
  theme_minimal() +
  theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1)) +
  xlab("Año") +
  ylab("Número de observaciones") 

g_year
# En 2007 no se han generado observaciones positivas
# Causa: En el shapefile que contiene los incendios producidos en 2007 solo hay 4 observaciones y ninguna tenía la fecha, por lo que estas no han podido utilizarse

# Valores muy bajos en 2008 y 2010, revisar
# 2010: correcto, en el archivo con los polígonos de incendios en 2010 solo hay 26 observaciones
# 2008: en el archivo relativo a ese año hay 37 observaciones, pero 11 de ellas no tienen la fecha de inicio, por lo que no se han podido utilizar
```


#### Día
```{r}
g_dia <- datos %>% 
  st_drop_geometry() %>% 
  # filter(fire==1) %>% 
  count(weekdays(date),fire) %>% 
  rename("dia" = "weekdays(date)") %>% 
  ggplot(aes(x = dia,y = n,fill=fire)) +
  geom_col(position="dodge",alpha=0.8) +
  scale_x_discrete(limits=c("lunes", "martes", "miércoles", "jueves", "viernes", "sábado", "domingo")) +
  theme(axis.text.x = element_text(angle = 45, vjust = 0.5, hjust=0.5)) +
 # geom_smooth(se=F) +
  scale_fill_hue(direction = -1) +
  theme_minimal() +
  xlab("Día") +
  ylab("Número de observaciones")

g_dia
```
Podría dar la impresión de que más incendios ocurren en fin de semana

```{r}
g = ggarrange(g_dia,g_mes,g_year,ncol=1,common.legend = T,legend = "right")

annotate_figure(g,top=text_grob("Distribución temporal de fire",size=18))
```

### Distribución espacial
```{r}
and <- esp_get_ccaa(ccaa = "Andalucía") %>% st_transform(st_crs(datos))
prov <- esp_get_prov() %>% filter(nuts2.name=="Andalucía") %>% st_transform(st_crs(datos))

# Casos positivos
g1 = ggplot(data = prov) + 
  geom_sf() +
  geom_sf(data = datos %>% filter(fire==1),size = 1, alpha = 0.4,col="red")+
  ggtitle("Casos positivos") +
  theme_minimal()

# Casos negativos
g0 = ggplot(data = prov) + 
  geom_sf() +
  geom_sf(data = datos %>% filter(fire==0),size = 1, alpha = 0.4,col="blue") +
  ggtitle("Casos negativos") +
  theme_minimal()

g = ggarrange(g1,g0,nrow=2)
annotate_figure(g,top=text_grob("Distribución espacial de fire",size=19))
```


## Análisis univariantes

### T2M


```{r}
ggplot(data = prov) + 
  geom_sf() +
  geom_sf(data = datos, aes(color=T2M),size = 1.2, alpha = 0.6) +
  facet_wrap(~month(date,label=TRUE)) + 
  scale_color_gradientn(colours = rainbow(5,rev=T)) +
  # scale_color_gradient(low="blue", high="red")+
  ggtitle("Distribución espacial de T2M por mes") + 
  theme_minimal() +
  theme(axis.text.x=element_blank(),
        axis.ticks.x=element_blank(),
        axis.text.y=element_blank(),
        axis.ticks.y=element_blank(),
        panel.grid.major = element_blank(),
        panel.grid.minor = element_blank())
```

Las temperaturas más elevadas se encuentran en el interior.

```{r}
datos %>% 
  st_drop_geometry() %>% 
  group_by(month(date,label=T),fire) %>% 
  summarize(AVG_T2M = mean(T2M),
            Mes = `month(date, label = T)`) %>% 
  ggplot(aes(x=Mes,y=AVG_T2M,fill=fire)) +
  geom_col(position="dodge",alpha=0.8) +
  scale_fill_hue(direction = -1) +
  theme_minimal()  
  # + theme(axis.text.x = element_text(angle = 45, vjust = 0.5, hjust=0.5))
```

En prácticamente todos los meses, la media de temperaturas en las observaciones positivas es mayor que en las observaciones negativas. Las temperaturas son más altas en los meses de verano, como cabía esperar.




### RH2M
```{r}
ggplot(data = prov) + 
  geom_sf() +
  geom_sf(data = datos, aes(color=RH2M),size = 1.2, alpha = 0.6) +
  facet_wrap(~month(date,label=TRUE)) + 
  scale_color_gradientn(colours = rainbow(5,rev=F)) +
  # scale_color_gradient(low="blue", high="red")+
  ggtitle("Distribución espacial de RH2M por mes") +
  theme_minimal() +
  theme(axis.text.x=element_blank(),
        axis.ticks.x=element_blank(),
        axis.text.y=element_blank(),
        axis.ticks.y=element_blank(),
        panel.grid.major = element_blank(),
        panel.grid.minor = element_blank())
```

La humedad del aire es mayor en las de costa. Las zonas más secas se encuentran en las zonas de interior del norte de Andalucía.

```{r}
datos %>% 
  st_drop_geometry() %>% 
  group_by(month(date,label=T),fire) %>% 
  summarize(AVG_RH2M = mean(RH2M),
            Mes = `month(date, label = T)`) %>% 
  ggplot(aes(x=Mes,y=AVG_RH2M,fill=fire)) +
  geom_col(position="dodge",alpha=0.8) +
  scale_fill_hue(direction = -1) +
  theme_minimal()  
  # + theme(axis.text.x = element_text(angle = 45, vjust = 0.5, hjust=0.5))
```

En prácticamente todos los meses, la humedad del aire media entre las observaciones positivas es menor que en las observaciones negativas. La humedad del aire disminuye significativamente en verano.

### GWETTOP

```{r}
ggplot(data = prov) + 
  geom_sf() +
  geom_sf(data = datos, aes(color=GWETTOP),size = 1, alpha = 0.6) +
  facet_wrap(~month(date,label=TRUE)) + 
  scale_color_gradientn(colours = rainbow(5,rev=F)) +
  # scale_color_gradient(low="blue", high="red")+
  ggtitle("Distribución espacial de GWETTOP por mes") +
  theme_minimal() +
  theme(axis.text.x=element_blank(),
        axis.ticks.x=element_blank(),
        axis.text.y=element_blank(),
        axis.ticks.y=element_blank(),
        panel.grid.major = element_blank(),
        panel.grid.minor = element_blank())
```

Las zonas con mayor humedad superficial de suelo son las que se encuentran en la costa mediterránea. Las zonas con un suelo más seco son las que se encuentran al norte de la cuenca el Guadalquivir, destacando la provincia de Huelva (hay que tener en cuenta que también es una zona en la que hay muchos incendios, la mayoría de los cuales se produce en verano, por lo que hay una gran concentración de observaciones durante el periodo estival, lo que explica el hecho de que se observen tantas observaciones con valores tan bajos de la humedad superficial de suelo).


```{r}
datos %>% 
  st_drop_geometry() %>% 
  group_by(month(date,label=T),fire) %>% 
  summarize(AVG_GWETTOP = mean(GWETTOP),
            Mes = `month(date, label = T)`) %>% 
  ggplot(aes(x=Mes,y=AVG_GWETTOP,fill=fire)) +
  geom_col(position="dodge",alpha=0.8) +
  scale_fill_hue(direction = -1) +
  theme_minimal()  
```

En general, la humedad superficial de suelo media entre los casos positivos es inferior que entre los casos negativos. Esto es especialmente evidendente entre los meses de primavera y otoño. En los meses de invierno esto no está tan claro, incluso parece revertirse la tendencia.

### WS10M
```{r}
ggplot(data = prov) + 
  geom_sf() +
  geom_sf(data = datos, aes(color=WS10M,alpha=WS10M,size=as.numeric(ifelse(WS10M<7.5,1,1.2)))) +
  scale_color_gradientn(colours = rainbow(5,rev=T)) +
  facet_wrap(~month(date,label=TRUE))+
  # scale_color_gradient(low="blue", high="red")+
  ggtitle("Distribución espacial de WS10M por mes") +
  theme_minimal() +
  theme(axis.text.x=element_blank(),
        axis.ticks.x=element_blank(),
        axis.text.y=element_blank(),
        axis.ticks.y=element_blank(),
        panel.grid.major = element_blank(),
        panel.grid.minor = element_blank()) +
  guides(size="none",alpha="none")
```

Los valores más elevados de la velocidad del viento a 10m de altura se dan en las zonas costeras.


```{r}
datos %>% 
  st_drop_geometry() %>% 
  group_by(month(date,label=T),fire) %>% 
  summarize(AVG_WS10M = mean(WS10M),
            Mes = `month(date, label = T)`) %>% 
  ggplot(aes(x=Mes,y=AVG_WS10M,fill=fire)) +
  geom_col(position="dodge",alpha=0.8)+
  scale_fill_hue(direction = -1) +
  theme_minimal()  
```

En todos los meses, los valores medios de la velocidad del viento a 10m sobre la superficie son significativamente mayores en las observaciones positivas.

### WD10M
```{r}
ggplot(data = prov) + 
  geom_sf() +
  geom_sf(data = datos, aes(color=WD10M),size = 1.2, alpha = 0.6) +
  facet_wrap(~month(date,label=TRUE))+ 
  # scale_color_stepsn(colours = rainborainbpw(8,rev=T)) +
  # scale_color_gradient(low="blue", high="red")+
  ggtitle("Distribución espacial de WD10M por mes") +
  theme_minimal() +
  theme(axis.text.x=element_blank(),
        axis.ticks.x=element_blank(),
        axis.text.y=element_blank(),
        axis.ticks.y=element_blank(),
        panel.grid.major = element_blank(),
        panel.grid.minor = element_blank()) +
  guides(color = guide_legend(override.aes = list(size = 5))) +
  scale_color_viridis_d(option="turbo")

```

```{r}
# Quisiera tener la dirección mayoritaria por mes y si hay incendio o no 

mode <- function(x) { names(which.max(table(x))) }

datos_mode_WD10M = datos %>%
  st_drop_geometry() %>%
  group_by(month(date,label=T),fire) %>%
  summarize(mode_WD10M = mode(WD10M)) 

tibble(mes = datos_mode_WD10M[datos_mode_WD10M$fire==1,1],
       fire1 = datos_mode_WD10M[datos_mode_WD10M$fire==1,3],
       fire0 = datos_mode_WD10M[datos_mode_WD10M$fire==0,3]) 

# En realidad de esto creo que no saco nada
```


### PRECTOTCORR

```{r}
ggplot(data = prov) + 
  geom_sf() +
  geom_sf(data = datos, aes(color=PRECTOTCORR,size=PRECTOTCORR,alpha=PRECTOTCORR)) +
  facet_wrap(~month(date,label=TRUE)) + 
  scale_color_gradientn(colours = rainbow(5,rev=F)) +
  guides(size="none") +
  # scale_color_gradient(low="blue", high="red")+
  ggtitle("Distribución espacial de PRECTOTCORR por mes") +
  theme_minimal() +
  theme(axis.text.x=element_blank(),
        axis.ticks.x=element_blank(),
        axis.text.y=element_blank(),
        axis.ticks.y=element_blank(),
        panel.grid.major = element_blank(),
        panel.grid.minor = element_blank()) +
  guides(alpha="none")
```

En la gran mayoría de las observaciones no se han observado precipitaciones. Se observan algunos valores positivos de precipitaciones distribuidos por el territorio Andaluz, alcanzándose los máximos en observaciones localizadas ne las coordilleras béticas.

```{r}
datos %>% 
  st_drop_geometry() %>% 
  group_by(month(date,label=T),fire) %>% 
  summarize(AVG_PRECTOTCORR = mean(PRECTOTCORR),
            Mes = `month(date, label = T)`) %>% 
  ggplot(aes(x=Mes,y=AVG_PRECTOTCORR,fill=fire)) +
  geom_col(position="dodge",alpha=0.8) +
  scale_fill_hue(direction = -1) +
  theme_minimal()  
```

Se observa una clara diferencia en la media mensual de precipitaciones en ambas clases en todos los meses, siendo mucho mayores entre las observaciones negativas. En los meses de verano esto sigue siendo cierto, si bien las diferencias se suavizan significativamente, ya que en ambos clases las precipitaciones son reducidas.


 <!-- IDEA PARA PODER EVALUAR MEJOR EL RENDIMIENTO REAL DEL MODELO: ESTRATIFICAR POR MESES -->
<!-- Ya que el objetivo no es que el modelo sea capaz de clasificar muy bien los datos del dataset, si no su utilidad práctica de cara predecir incendios forestales. -->
<!-- EL objetivo es que la muestra permita detectar los patrones que causan los incendios forestales, por lo que el objetivo al generar las muestras negativas es evitar en la mayor medida posible los sesgos debidos al proceso de selección de la muestra. Se busca que el modelo sea capaz de discriminar las observaciones en riesgo de incendio forestal, no los sesgos introducidos en el proceso de generación de la muestra. -->

<!-- Al generar los ejemplos negativos, mi objetivo es intentar evitar los sesgos que se puedan producir en el proceso de seleccion, con el objetivo de que la diferencia entre ejemplos positivos y negativos sea realmente debida a las carácterísticas de los incendios y no a sesgos introducidos en el proceso de muestreo. -->

### NDVI
```{r}
ggplot(data = prov) + 
  geom_sf() +
  geom_sf(data = datos, aes(color=NDVI),size = 1.2, alpha = 0.6) +
  facet_wrap(~month(date,label=TRUE)) + 
  scale_color_gradientn(colours = rainbow(5,rev=T)) +
  # scale_color_gradient(low="blue", high="red")+
  ggtitle("Distribución espacial de NDVI por mes") +
  theme_minimal() +
  theme(axis.text.x=element_blank(),
        axis.ticks.x=element_blank(),
        axis.text.y=element_blank(),
        axis.ticks.y=element_blank(),
        panel.grid.major = element_blank(),
        panel.grid.minor = element_blank())
```

Los valores más elevados del NDVI se encuentran en el norte de la comunidad (Sierra Morena y Sierra de Aracena) y en la zona sur-centro. Los valores son especialmente elevados en la Sierra de Grazalema y en las marismas del Guadalquivir y especialmente bajos en el este de la comunidad (Almería).

```{r}
datos %>% 
  st_drop_geometry() %>% 
  group_by(month(date,label=T),fire) %>% 
  summarize(AVG_NDVI = mean(NDVI),
            Mes = `month(date, label = T)`) %>% 
  ggplot(aes(x=Mes,y=AVG_NDVI,fill=fire)) +
  geom_col(position="dodge",alpha=0.8) +
  scale_fill_hue(direction = -1) +
  theme_minimal()    
  # theme(axis.text.x = element_text(angle = 45, vjust = 0.5, hjust=0.5))
```

El valor del NDVI disminuye en los meses de verano y alcanza su máximo en los meses de invierno. No se observa claramente una relación entre el NDVI y la clase de la observación en la muestra.

### poblacion

```{r}
ggplot(data = prov) + 
  geom_sf() +
  geom_sf(data = datos, aes(color=poblacion,size = poblacion,alpha=poblacion/median(poblacion))) +
  scale_color_gradientn(colours = rainbow(5,rev=T)) +
  guides(size="none",alpha="none") +
  # scale_color_gradient(low="blue", high="red")+
  ggtitle("Distribución espacial de poblacion") + 
  theme_minimal() #+
  # theme(axis.text.x=element_blank(),
  #       axis.ticks.x=element_blank(),
  #       axis.text.y=element_blank(),
  #       axis.ticks.y=element_blank(),
  #       panel.grid.major = element_blank(),
  #       panel.grid.minor = element_blank())
```

En la gran mayoría del territorio se observan niveles de población inferiores a 20.000 habitantes. Destacan las capitales de provincia por tener valores significativamente más elevados.

```{r}
# Este en realidad no sirve para nada, pero es curioso

datos %>% 
  st_drop_geometry() %>% 
  group_by(month(date,label=T),fire) %>% 
  summarize(AVG_poblacion = mean(poblacion),
            Mes = `month(date, label = T)`) %>% 
  ggplot(aes(x=Mes,y=AVG_poblacion,fill=fire)) +
  geom_col(position="dodge",alpha=0.8)
```

```{r}
by(datos$poblacion,datos$fire,summary)
```


### dens_poblacion
```{r}
ggplot(data = prov) + 
  geom_sf() +
  geom_sf(data = datos, aes(color=dens_poblacion, size= dens_poblacion, alpha=dens_poblacion)) +
# facet_wrap(~fire)+
# theme(axis.text.x=element_blank(),
#       axis.ticks.x=element_blank(),
#       axis.text.y=element_blank(),
#       axis.ticks.y=element_blank()) + 
  scale_color_gradientn(colours = rainbow(5,rev=T)) +
  guides(size="none",alpha="none") +
  # scale_color_gradient(low="blue", high="red")+
  ggtitle("Distribución espacial de dens_poblacion") +
  theme_minimal()
```

Comentario similar a la población. Mayor densidad de población en las zonas de costa y en las capitales de provincia.

```{r}
by(datos$dens_poblacion,datos$fire,summary)
```

### pendiente
```{r}
ggplot(data = prov) + 
  geom_sf() +
  geom_sf(data = datos, aes(color=pendiente, size= pendiente, alpha=pendiente)) +
    scale_color_gradientn(colours = rainbow(5,rev=T)) +
  guides(size="none",alpha="none") +
  ggtitle("Distribución espacial de pendiente") +
  theme_minimal()
```


### elevacion
```{r}
ggplot(data = prov) + 
  geom_sf() +
  geom_sf(data = datos, aes(color=elevacion, size= elevacion, alpha=elevacion)) +
    scale_color_gradientn(colours = rainbow(5,rev=T)) +
  guides(size="none",alpha="none") +
  ggtitle("Distribución espacial de elevacion") +
  theme_minimal()
```

### curvatura
```{r}
ggplot(data = prov) + 
  geom_sf() +
  geom_sf(data = datos, aes(color=curvatura, alpha=curvatura)) +
    scale_color_gradientn(colours = rainbow(5,rev=T)) +
  guides(size="none",alpha="none") +
  ggtitle("Distribución espacial de curvatura") +
  theme_minimal()
```



### dist_carretera
```{r}
ggplot(data = prov) + 
  geom_sf() +
  geom_sf(data = datos, aes(color=dist_carretera, alpha=dist_carretera)) +
    scale_color_gradientn(colours = rainbow(5,rev=T)) +
  guides(size="none",alpha="none") +
  ggtitle("Distribución espacial de dist_carretera") +
  theme_minimal()
```


### dist_electr
```{r}
ggplot(data = prov) + 
  geom_sf() +
  geom_sf(data = datos, aes(color=dist_electr, alpha=dist_electr)) +
    scale_color_gradientn(colours = rainbow(5,rev=T)) +
  guides(size="none",alpha="none") +
  ggtitle("Distribución espacial de dist_electr") +
  theme_minimal()
```

### dist_camino
```{r}
ggplot(data = prov) + 
  geom_sf() +
  geom_sf(data = datos, aes(color=dist_camino, alpha=dist_camino)) +
    scale_color_gradientn(colours = rainbow(5,rev=T)) +
  guides(size="none",alpha="none") +
  ggtitle("Distribución espacial de dist_camino") +
  theme_minimal()
```

### dist_sendero
```{r}
ggplot(data = prov) + 
  geom_sf() +
  geom_sf(data = datos, aes(color=dist_sendero, alpha=dist_sendero)) +
    scale_color_gradientn(colours = rainbow(5,rev=T)) +
  guides(size="none",alpha="none") +
  ggtitle("Distribución espacial de dist_sendero") +
  theme_minimal()
```


### dist_poblacion
```{r}
ggplot(data = prov) + 
  geom_sf() +
  geom_sf(data = datos, aes(color=dist_poblacion, alpha=dist_poblacion)) +
    scale_color_gradientn(colours = rainbow(5,rev=T)) +
  guides(size="none",alpha="none") +
  ggtitle("Distribución espacial de dist_poblacion") +
  theme_minimal()
```

### dist_ferrocarril
```{r}
ggplot(data = prov) + 
  geom_sf() +
  geom_sf(data = datos, aes(color=dist_ferrocarril, alpha=dist_ferrocarril)) +
    scale_color_gradientn(colours = rainbow(5,rev=T)) +
  guides(size="none",alpha="none") +
  ggtitle("Distribución espacial de dist_ferrocarril") +
  theme_minimal()
```

### dist_rios
```{r}
ggplot(data = prov) + 
  geom_sf() +
  geom_sf(data = datos, aes(color=dist_rios, alpha=dist_rios)) +
    scale_color_gradientn(colours = rainbow(5,rev=T)) +
  guides(size="none",alpha="none") +
  ggtitle("Distribución espacial de dist_rios") +
  labs(color = "Distancia a río") +
  theme_minimal() +
  theme(legend.position = "bottom",legend.key.width = unit(3,"line"),legend.title.position = "top")
```


## Variables numéricas

En los datos hay 18 variables numéricas.

```{r}
datos_numeric = datos |> select(where(is.numeric)) %>% st_drop_geometry()

# datos_numeric$fire = as.numeric(datos$fire)
```


### Correlaciones:

Se estudian primero las correlaciones entre las variables.

```{r}
R = cor(datos_numeric)

corrplot(R,method = "ellipse",type = "lower")

summary(R-diag(diag(R)))
```

Las variables más correlacionadas en la muestra son T2M con RH2M (negativamente, -0.71), T2M con GWETTOP (negativamente, -0.69) y GWETTOP con R2HM (positivamente, 0.68). Tiene sentido que la humedad del aire a dos metros esté correlacionada positivamente con la humedad del suelo y que ambas estén -egativamente correlacionadas con la temperatura del aire. También están correladas la población con la densidad de población (positivamente,0.63).



### Gráfico de coordenadas paralelas

```{r}
datos |> 
  select(fire,where(is.numeric)) |> 
  ggparcoord(columns=2:19,alphaLines=0.1,groupColumn = "fire") +
  xlab('') + 
  ylab('') + 
  scale_x_discrete(labels=colnames(datos_numeric)) +
  scale_color_hue(direction = -1) +
  # guides(color ="none")+
  theme_minimal() + 
  ggtitle("Tipificación a normal estándar (por defecto)") +
  theme(plot.title = element_text(size = 9),
        axis.text.x = element_text(angle = 90))
```


Considerando las variables tipificadas, se observa que aquellas con más variabilidad son PRECTOTCORR, población y dens_poblacion. Todas estas variables tienen un sesgo positivo, lo cual es razonable pues todas ellas son variables que solo toman valores positivos.


### Box-plots

```{r}
boxplots <- map(names(datos_numeric), ~ ggplot(datos, aes(x = fire, y = .data[[.x]],fill=fire)) +
                  geom_boxplot() +
                  scale_fill_hue(direction = -1) +
                  guides(fill="none") +
                  labs(title = paste(.x, "~ fire")))

ggarrange(plotlist = boxplots,ncol=6,nrow=3)
# do.call(ggarrange,c(boxplots,ncol = 3,nrow=6))
```

Cabe destacar por un lado la diferencia de escala entre las variables y por otro la gran cantidad de valores atípicos presentes en los datos. Es importante destacar que los valores atípicos no son erróneos, son inherentes a la naturaleza de los datos. Por ejemplo, veamos el caso de PRECTOTCORR.

Se puede que esta variable presenta valores atípicamente altos, que se dan en observaciones en las que no se ha observado incendio. 

```{r}
by(datos$PRECTOTCORR,datos$fire,summary)
```

Acumular 46.06mm de lluvia en un día es un valor posible, probablemente atípico en esta región pero no hay motivo para considerar que es un valor erróneo. Es coherente que las observaciones en las que se han registrado mayores niveles de precipitación sean observaciones negativas.


<!-- ### Matriz de dispersion -->
<!-- ```{r} -->
<!-- # Inviable dada la cantidad de datos -->
<!-- datos_numeric |>   -->
<!--   mutate(fire = datos$fire) |>  -->
<!--   ggpairs(aes(color=fire)) -->
<!-- ``` -->



## Variables categóricas

Las variables date, cod_municipio y municipio son "identificadoras". Las variables explicativas categóricas son WD10M, orientacion, enp y uso_suelo. 


```{r}
datos_categ <- datos |> select(WD10M,orientacion,enp,uso_suelo) %>% st_drop_geometry()
```

### Histogramas

```{r}
histogramas <- map(names(datos_categ), ~ ggplot(datos, aes(x = .data[[.x]], fill = fire)) +
                    geom_bar(position = "dodge", alpha = 0.8) +
                    scale_fill_hue(direction = - 1) +
                    theme_minimal() +
                    labs(title = .x)+theme(axis.text.x = element_text(size=7)))

ggarrange(plotlist = histogramas, ncol = 2, nrow = ceiling(length(histogramas)/2))


uso_suelo = datos |> 
  ggplot(aes(x = uso_suelo,fill=fire)) +
  geom_bar(position = "dodge", alpha = 0.8) +
  scale_fill_hue(direction = - 1) +
  theme_minimal() +
  theme(axis.text.x = element_text(size=7))
```

En el caso de la variable `WD10M` cabe comentar la escasez de observaciones con dirección norte. En el gráfico no está clara la relación de esta variable con la variable objetivo.

En el caso de la variable `orientacion` la relación tampoco está clara, aunque parece que en las observaciones orientadas al sur (SE,S,SW) es más común observar incendios.

La variable `enp` no parece mostrar diferencias en ambas clases.

En el caso de la variable `uso_suelo`, sí que se observan diferencias significativas en ambas clases, que sin duda también son debidas al proceso de muestreo seguido.

```{r}
datos |> 
  mutate(uso_suelo = fct_lump(uso_suelo,
                              n = 7, 
                              other_level= "Otro"))  |>
  ggplot() + 
  geom_sf(data = prov) + 
  geom_sf(aes(color=uso_suelo),size = 1.2, alpha = 0.6) +
  ggtitle("Distribución espacial de uso_suelo") +
  theme_minimal() +
  guides(color = guide_legend(override.aes = list(size = 5))) +
  scale_color_viridis_d(option="turbo")
  # scale_color_stepsn(colours = rainborainbpw(8,rev=T)) +
  # scale_color_gradient(low="blue", high="red")+

   

```


## PCA

### Solo variables numericas
```{r}
pca <- princomp(datos_numeric,cor=T)

library(knitr)
summary(pca) 
str(s)
pca$loadings
```

Se necesitan al menos 11 componentes principales para explicar el 80% de la variabilidad de las 18 variables numéricas de la muestra (tipificadas) y al menos 14 para explicar el 90%. Esto refleja, la complejidad del conjunto de datos.

## Cluster

### Variables numéricas
Se aplica un algoritmo de clustering basado en k-medias adaptado para grandes conjuntos de datos.

```{r}
library(cluster)
clara<- clara(datos_numeric,2)
table(clara$clustering)

# Si los datos no están normalizados, el método no es capaz de separar los datos
```

Vemos que la partición obtenida con los datos a penas consigue separarlos. Se debe a que algunas variables tienen mucha más variabilidad que otras y el algoritmo de k-medias no funciona bien cuando las variables tienen varianzas muy diferentes. 

Probamos a normalizarlos antes de aplicar el algoritmo.

```{r}
set.seed(12345)
# Probamos tipificando los datos:
clara = datos_numeric |> 
  scale() |> 
  clara(k=2)

table(clara$clustering)
# Ahora el algoritmo consigue separar mejor los datos

t = table(clara$clustering,datos$fire)
t = rbind(t[2,],t[1,])
rownames(t) <- paste0("cluster_",c(1,2))
colnames(t) <- c("No fire","Fire")
t
# addmargins(t)

prop.table(t,2)*100

sum(diag(t))/nrow(datos) # Proporción de observaciones "correctamente separada"
```

Al normalizar previamente los datos, sí se consigue una partición más balanceada de la muestra. Sin embargo, esta no parece aportar ningún tipo de información sobre la variable objetivo.


<!-- ### Todas las variables -->

<!-- ```{r} -->
<!-- # Probamos tipificando los datos: -->
<!-- clara = M |>  -->
<!--   scale() |>  -->
<!--   clara(k=2) -->

<!-- table(clara$clustering) -->
<!-- # Ahora el algoritmo consigue separar mejor los datos -->

<!-- t = table(clara$clustering,datos$fire) -->
<!-- rownames(t) <- paste0("cluster_",c(1,2)) -->
<!-- colnames(t) <- c("No fire","Fire") -->
<!-- t -->
<!-- # addmargins(t) -->

<!-- prop.table(t,2)*100 -->

<!-- sum(diag(t))/nrow(datos) # Proporción de observaciones "correctamente separada" -->
<!-- ``` -->

<!-- No parece aportar nada en el análisis. -->
