==> scripts/strat/6_casos_interes.Rmd <==
---
title: "Casos de interés"
author: "Juan Baeza Ruiz-Henestrosa"
date: "`r Sys.Date()`"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
knitr::opts_knit$set(root.dir = "D:/usuario/Documents/Universidad/5º/TFG - organizado") # Para que haga knitr desde el directorio del proyecto

```


```{r}
library(tidyverse)
library(tidymodels)
library(sf)
library(ggplot2)
library(terra)
library(mapSpain)
library(skimr)
library(ggpubr)
```

```{r}
load("salidas_intermedias/datos_strat_depurados_geom_2024-04-27.RData") # Datos
datos <- datos |> 
  mutate(uso_suelo = fct_lump(uso_suelo,
                              n = 7, # nos quedamos con los 7 niveles del factor más frecuentes (clases 2 y 3)
                              other_level= "Otro"))
```


<!-- Para poder estudiar casos concretos de interés, sin tener que hacer frente a todas las complicaciones derivadas del trabajar con datos raster, un enfoque puede ser el siguiente: -->


<!-- 1. Estudiar un incendio concreto un día y ver como evoluciona el riesgo de incendio en los días cercanos, en el área de interés (un área que contenga al incendio que se quiere estudiar). -->

<!-- 2. Generar puntos aleatoriamente en el área de interés (o formando una cuadrícula) en cada uno de los días cercanos al incendio.  -->

<!-- 3. Asociar a cada punto sus variables correspondientes siguiendo el procedimiento usado para generar la muestra. -->

<!-- 4. Usar el modelo para clasificar cada punto (o predecir la probabilidad de incendio). -->

<!-- 5. Mostrar los gráficos de cada día con el polígono del incendio y compararlos para estudiar la capacidad de discretización del modelo. -->

<!-- 6. En función del rendimiento del modelo se puede probar a extender el área de estudio y usar instantes de tiempo más alejados entre sí (p.e. semanas), pues es razonable que en días próximos y localizaciones cercanas los riesgos no varíen en exceso. -->

<!-- Si se obtienen resultados razonables, puede probarse a añadir variaciones en las variables climáticas (p.e. temperatura) para ver sus efectos. -->


## Visión general del desempeño del modelo


```{r}
and <- esp_get_ccaa(ccaa = "Andalucía") %>% st_transform(st_crs(datos))
```

Se construye la rejilla
```{r}
# grid de puntos 10km x 10km de andalucía
grid = st_make_grid(and,
                    cellsize = c(10000,10000), # 10000
                    what = "centers")[and]

g = and %>% 
  ggplot() +
  geom_sf() +
  theme_minimal()
 
g + 
  geom_sf(data = grid,color="red",size=0.6) +
  labs(title = "Malla de puntos con una resulción de 10km x 10km")
```

Se va a analizar cada observación el 15 de cada mes.
```{r}
sample = NULL

for (m in 1:12) {
  if (is.null(sample)){
    sample <- tibble(date = rep(ymd(paste("2022",m,"15",sep="/"))),geometry = grid) %>% st_sf()
  } else
  sample <- sample %>% 
    bind_rows(tibble(date = rep(ymd(paste("2022",m,"15",sep="/"))),geometry = grid))
}

```

Se le asocian todas las variables correspondientes a cada observación.
```{r}
source("scripts/strat/fun_asignar_variables.R")
full_grid  = asignar_variables(sample)

full_grid
```


Se imputan los valores faltantes de NDVI asignandoles los del año anterior, ya que faltan los archivos de marzo y diciembre de 2022.
```{r}
# La siguiente función lee el archivo con el NDVI correspondiente a un mes y a un año dados (si está disponible)

read_NDVI = function(MM,YYYY) {
  MM = str_pad(as.character(MM),2,"left",pad = "0")
  YY = substr(as.character(YYYY),3,4)
  if (as.numeric(YY)<=06) {
    ruta <- paste0("data_raw/vegetacion/",YYYY,"TERMODMEDMNDVI/InfGeografica/InfRaster/TIFF/TERMOD_",YY,MM,"01_h17v05_medmndvi.tif")
  } else if (as.numeric(YY)<=11) {
    ruta <- paste0("data_raw/vegetacion/",YYYY,"TERMODMEDMNDVI/InfGeografica/InfRaster/TIF/TERMOD_",YY,MM,"01_h17v05_medmndvi.tif")
  } else if (as.numeric(YY)<=21){
    ruta <- paste0("data_raw/vegetacion/",YYYY,"TERMODMEDMNDVI/InfGeografica/InfRaster/TIFF/termod_",YY,MM,"01_h17v05_medmndvi.tif")
  }else {
    ruta <- paste0("data_raw/vegetacion/",YYYY,"TERMODMEDMNDVI/InfGeografica/InfRaster/COG/termod_",YY,MM,"01_h17v05_medmndvi_COG.tif")
  }

  if (file.exists(ruta)) {
    NDVI = rast(ruta)
  } else
    NDVI = NA
  return(NDVI)
}

# Los meses para los cuales no está disponible el archivo de NDVI:
year_month_missing_NDVI = c(
"2003-01",
"2003-04",
"2017-02",
"2018-11",
"2020-11",
"2021-12",
"2022-03",
"2022-12")

# Para cada observación para la que no está disponible el NDVI (porque la información de ese mes no está disponible), se obtiene el correspondiente al mismo mes del año anterior, si este está disponible, si no, el del año posterior:

missing_NDVI = full_grid |> 
  filter(is.na(NDVI)) |> 
  mutate(year = year(date),month = month(date)) |> 
  group_by(year,month) |> 
  filter(paste(year,str_pad(as.character(month),2,"left",pad = "0"),sep="-") %in%   year_month_missing_NDVI) |>  # Se filtra porque también hay observaciones que tienen NA en el NDVI no porque no exista el archivo, si no lo mismo que en ocasiones anteriores, porques discrepancias entre los límites de los polígonos
  nest() |>    
  mutate(NDVI_rast = map2(month,year-1,read_NDVI), # Se lee primero el del año anterior
         NDVI_rast = ifelse(is.na(unlist(NDVI_rast)),map2(month,year+1,read_NDVI),NDVI_rast), # Si no está disponible, se toma el del año posterior
         NDVI_nuevo = map2(NDVI_rast,data,~terra::extract(.x,.y)[,2])) |> 
  select(-NDVI_rast) |> 
  unnest(c(data,NDVI_nuevo)) |> 
  mutate(NDVI = NDVI_nuevo,.keep="unused")

ind_modificados = is.na(full_grid$NDVI) & (paste(year(full_grid$date),str_pad(as.character(month(full_grid$date)),2,"left",pad = "0"),sep="-") %in% year_month_missing_NDVI) # Los elementos que han sido modificados

# Se asignan los valores imputados del NDVI:
full_grid[ind_modificados,]$NDVI = missing_NDVI$NDVI
```

```{r}
summary(full_grid)

datos %>%  st_drop_geometry %>% skim(.data_name = "datos")
```

Se agrupan los niveles de uso de suelo
```{r}
full_grid <- full_grid |> 
  mutate(uso_suelo = fct_other(uso_suelo,
                               keep = c("21","22","23","24","31","32","33"),
                               other_level= "Otro"))

full_grid %>% st_drop_geometry() %>% skim()
# save(full_grid,file = "salidas_intermedias/full_grid_meses_2022_processed.RData")
load("salidas_intermedias/full_grid_meses_2022_processed.RData")
```

Se usará el modelo de regresión logística lasso final,
```{r}
load("Private/all_models_test.RData")

model <- models %>% filter(model_name=="lr")
# model <- models %>% filter(model_name=="svm_linear")
# model <- models %>% filter(model_name=="rf")

rm(models) # Es un archivo muy pesado, lo eliminamos de la memoria
```


```{r}
pred_class = model %>% 
  pull(last_fit) %>% 
  .[[1]] %>% 
  extract_workflow() %>% 
  predict(new_data = full_grid)
pred_probs = model %>% 
  pull(last_fit) %>% 
  .[[1]] %>% 
  extract_workflow() %>% 
  predict(new_data = full_grid,type="prob")
pred = cbind(full_grid,pred_class,pred_probs)
```

Salidas gráficas:
```{r}
g <- ggplot(data = and) + 
  geom_sf() +
  geom_sf(data = pred, aes(color=.pred_1, alpha = .pred_1),size = 1.5) +
  facet_wrap(~month(date,label=TRUE)) + 
  scale_color_gradientn(colours = rainbow(5,rev=T),limits=c(0,1)) +
  # scale_color_gradient(low="blue", high="red")+
  labs(title="Probabilidad estimada de incendio el día 15 de cada mes de 2022",
       subtitle = paste0("Model: ",model$model_name)) + 
  guides(alpha = "none") + 
  theme_minimal() +
  theme(axis.text.x=element_blank(),
        axis.ticks.x=element_blank(),
        axis.text.y=element_blank(),
        axis.ticks.y=element_blank(),
        panel.grid.major = element_blank(),
        panel.grid.minor = element_blank())

g

# Hacer el gráfico con los mejores modelos
```

Le añadimos los incendios producidos en 2022
```{r}
incendios22 <- st_read(paste0("./data_raw/incendios_2000-2022/incendios_",2022,".shp")) %>% 
  st_transform(st_crs(full_grid)) %>% 
  mutate(date=ymd(fecha_inic))
```

```{r}
# OPCIÓN 1 (mejor)
ggplot(data = and) + 
  geom_sf() +
  geom_sf(data = pred, aes(color=.pred_1),alpha=0.8,size = 1.5) +
  facet_wrap(~month(date,label=TRUE)) + 
  scale_color_gradientn(colours = rainbow(5,rev=T),limits=c(0,1)) +
  # scale_color_gradient(low="blue", high="red")+
  guides(alpha = "none") + 
  theme_minimal() +
  theme(axis.text.x=element_blank(),
        axis.ticks.x=element_blank(),
        axis.text.y=element_blank(),
        axis.ticks.y=element_blank(),
        panel.grid.major = element_blank(),
        panel.grid.minor = element_blank()) +
  geom_sf(data = incendios22 %>% st_centroid, color = "black", shape =24, size=1,fill = "red") + 
  labs(title = "Probabilidad de incendio estimada el día 15 de cada mes de 2022",
       subtitle = paste0("Modelo: ",model$model_name),
       color = "Probalidad\nde incendio\nestimada")
```


```{r}
 ggplot(data = and) + 
  geom_sf() +
  geom_sf(data = pred, aes(color=.pred_1),alpha=0.5,size = 1.5) +
  facet_wrap(~month(date,label=TRUE)) + 
  scale_color_gradientn(colours = rainbow(5,rev=T),limits=c(0,1)) +
  # scale_color_gradient(low="blue", high="red")+
  ggtitle("Probabilidad estimada de incendio forestal + Incendios observados") + 
  guides(alpha = "none") + 
  theme_minimal() +
  theme(axis.text.x=element_blank(),
        axis.ticks.x=element_blank(),
        axis.text.y=element_blank(),
        axis.ticks.y=element_blank(),
        panel.grid.major = element_blank(),
        panel.grid.minor = element_blank()) +
  geom_sf(data = incendios22, color = "firebrick",fill="transparent") +
  geom_sf(data = st_buffer(incendios22,dist=5000), color = "firebrick",fill="firebrick",alpha=0.5,linetype="dotted")
  
```

Prácticamente todos los incendios se han producido en zonas con probabilidad de riesgo elevada. Es interesante como las zonas con alto riesgo de incendio van variando cada mes y, además, como se detectan zonas aisladas con una mayor probabilidad de incendio de incendio que se corresponde con las zonas en las que han ocurrido incendios. Ejemplo de esto son los meses de agosto, octubre o noviembre. En estos meses en los que, en general, la probabilidad estimada de incendio es significativamente menor a junio y julio se, se detectan zonas más o menos aisladas en las que la probabilidad de incendio es significativamente mayor. Y es en estas zonas, en general, en las que se han observado incendios.


```{r}
# ESTO EN REALIDAD ES UNA MALA IDEA PONERLO, PORQUE NO SE ESTÁN SELECCIONANDO LOS PUNTOS QUE DEBERÍAN

# En las localizaciones en las que ha habido incendio
pred[incendios22,]$.pred_1 %>% summary()
pred[incendios22,] %>% 
  ggplot(aes(y = .pred_1)) +
  geom_boxplot(fill = "grey") +
  theme_minimal() +
  labs(title = "Boxplot de la probabilidad estimada de incendio en las localizaciones en
       las que ha habido incendio")

pred[incendios22,]$.pred_1 %>% boxplot()
```

## Efecto del aumento de las temperaturas
```{r}
temp05 <- full_grid %>% 
  mutate(T2M = 1.05*T2M)

pred_class_temp05 = lr_final %>% 
  pull(last_fit) %>% 
  .[[1]] %>% 
  extract_workflow() %>% 
  predict(new_data = temp05)
pred_probs_temp05 = lr_final %>% 
  pull(last_fit) %>% 
  .[[1]] %>% 
  extract_workflow() %>% 
  predict(new_data = temp05,type="prob")
pred_temp05 = cbind(full_grid,pred_class_temp05,pred_probs_temp05)

g_temp05 <- ggplot(data = and) + 
  geom_sf() +
  geom_sf(data = pred_temp05, aes(color=.pred_1, alpha = .pred_1),size = 1.5) +
  facet_wrap(~month(date,label=TRUE)) + 
  scale_color_gradientn(colours = rainbow(5,rev=T),limits=c(0,1)) +
  # scale_color_gradient(low="blue", high="red")+
  ggtitle("temp + 5%") + 
  guides(alpha = "none") + 
  theme_minimal() +
  theme(axis.text.x=element_blank(),
        axis.ticks.x=element_blank(),
        axis.text.y=element_blank(),
        axis.ticks.y=element_blank(),
        panel.grid.major = element_blank(),
        panel.grid.minor = element_blank())

g_temp05
g
```


```{r}
temp10 <- full_grid %>% 
  mutate(T2M = 1.10*T2M)

pred_class_temp10 = lr_final %>% 
  pull(last_fit) %>% 
  .[[1]] %>% 
  extract_workflow() %>% 
  predict(new_data = temp10)
pred_probs_temp10 = lr_final %>% 
  pull(last_fit) %>% 
  .[[1]] %>% 
  extract_workflow() %>% 
  predict(new_data = temp10,type="prob")
pred_temp10 = cbind(full_grid,pred_class_temp10,pred_probs_temp10)

g_temp10 <- ggplot(data = and) + 
  geom_sf() +
  geom_sf(data = pred_temp10, aes(color=.pred_1, alpha = .pred_1),size = 1.5) +
  facet_wrap(~month(date,label=TRUE)) + 
  scale_color_gradientn(colours = rainbow(5,rev=T),limits=c(0,1)) +
  # scale_color_gradient(low="blue", high="red")+
  ggtitle("temp + 10%") + 
  guides(alpha = "none") + 
  theme_minimal() +
  theme(axis.text.x=element_blank(),
        axis.ticks.x=element_blank(),
        axis.text.y=element_blank(),
        axis.ticks.y=element_blank(),
        panel.grid.major = element_blank(),
        panel.grid.minor = element_blank())

g_temp10

g
```

```{r}
mean(pred$.pred_1)
mean(pred_temp05$.pred_1)
mean(pred_temp10$.pred_1)

boxplot(data.frame(Original = pred$.pred_1,
                   Temp05 =pred_temp05$.pred_1,
                   Temp10 =pred_temp10$.pred_1))
```


```{r}
clim1 <- full_grid %>% 
  mutate(T2M = 1.10*T2M,
         GWETTOP = 0.8*GWETTOP,
         PRECTOTCORR = 0.8*PRECTOTCORR,
         NDVI = 0.8*NDVI,
         RH2M = 0.8*RH2M)

pred_class_clim1 = lr_final %>% 
  pull(last_fit) %>% 
  .[[1]] %>% 
  extract_workflow() %>% 
  predict(new_data = clim1)
pred_probs_clim1 = lr_final %>% 
  pull(last_fit) %>% 
  .[[1]] %>% 
  extract_workflow() %>% 
  predict(new_data = clim1,type="prob")
pred_clim1 = cbind(full_grid,pred_class_clim1,pred_probs_clim1)

g_clim1 <- ggplot(data = and) + 
  geom_sf() +
  geom_sf(data = pred_clim1, aes(color=.pred_1, alpha = .pred_1),size = 1.5) +
  facet_wrap(~month(date,label=TRUE)) + 
  scale_color_gradientn(colours = rainbow(5,rev=T),limits=c(0,1)) +
  # scale_color_gradient(low="blue", high="red")+
  ggtitle("temp + 10%") + 
  guides(alpha = "none") + 
  theme_minimal() +
  theme(axis.text.x=element_blank(),
        axis.ticks.x=element_blank(),
        axis.text.y=element_blank(),
        axis.ticks.y=element_blank(),
        panel.grid.major = element_blank(),
        panel.grid.minor = element_blank())

g_clim1
g
```

## Caso de interés

Se va a intentar estudiar el incendio del 8 de septiembre de 2021 en Sierra Bermeja por ser de especial relevancia.


```{r}
incendios21 <- st_read(paste0("./data_raw/incendios_2000-2022/incendios_",2021,".shp")) %>% 
  st_transform(st_crs(datos)) %>% 
  mutate(FECHA_INIC=ymd(FECHA_INIC))

incendio_estudio <- incendios21 %>% filter(month(FECHA_INIC)==9)

and <- esp_get_ccaa(ccaa = "Andalucía") %>% st_transform(st_crs(datos))
rm(incendios21)
```


```{r}
prov <- esp_get_prov() %>% filter(nuts2.name=="Andalucía") %>% st_transform(st_crs(datos))

g = prov %>% 
  ggplot() +
  geom_sf() + 
  geom_sf(data = incendio_estudio, color="red",fill=alpha("red",0.2))+
  theme_minimal()+
  theme(axis.text.x=element_blank(),
        axis.ticks.x=element_blank(),
        axis.text.y=element_blank(),
        axis.ticks.y=element_blank(),
        panel.grid.major = element_blank(),
        panel.grid.minor = element_blank()) 
  
  # labs(title = "Caso de estudio: Incendio de Sierra Bermeja",
  #      subtitle = "8 de septiembre de 2021") +
  
  # theme(plot.title = element_text(face="bold"))
g
```


```{r}
bbox = st_buffer(incendio_estudio,dist=10000) %>% 
  st_bbox() 
#   as.matrix() %>% t() %>% 
#   as.data.frame()

# ggplot(bbox %>% st_as_sfc()) + 
#   geom_sf() + 
#   geom_sf(data = st_buffer(incendio_estudio,dist=10000)) +
#   geom_sf(data=incendio_estudio)

g1 = g + coord_sf(xlim=c(bbox$xmin,bbox$xmax),ylim=c(bbox$ymin,bbox$ymax))
 
g1 = prov %>% 
  ggplot() +
  geom_sf() + 
  geom_sf(data = incendio_estudio, color="red",fill=alpha("red",0.2))+
  theme_minimal() +
  # labs(title = "Caso de estudio: Incendio de Sierra Bermeja",
  #      subtitle = "8 de septiembre de 2021") +
  coord_sf(xlim=c(bbox$xmin,bbox$xmax),ylim=c(bbox$ymin,bbox$ymax))
  
gg <- ggarrange(g,g1)

annotate_figure(gg, top = text_grob("Caso de estudio: Incendio de Sierra Bermeja\n8 de septiembre de 2021", 
                                    size = 14))
```


```{r}
grid = st_make_grid(st_buffer(incendio_estudio,dist=10000),
                    cellsize = c(1000,1000),
                    what = "centers")[and]
g + 
  geom_sf(data = grid,size=0.7) +
  coord_sf(xlim=c(bbox$xmin,bbox$xmax),ylim=c(bbox$ymin,bbox$ymax))

```

```{r}
sample = tibble(date = rep(incendio_estudio$FECHA_INIC,length(grid)),geometry = grid) %>% # el día del incendio
                 bind_rows(tibble(date = rep(incendio_estudio$FECHA_INIC+15,length(grid)),geometry = grid)) %>%  # 15 días después
                 bind_rows(tibble(date = rep(incendio_estudio$FECHA_INIC+30,length(grid)),geometry = grid)) %>%  # el mes después
                 bind_rows(tibble(date = rep(incendio_estudio$FECHA_INIC+45,length(grid)),geometry = grid)) %>% # 45 días después del incendio
                bind_rows(tibble(date = rep(incendio_estudio$FECHA_INIC-15,length(grid)), geometry = grid)) %>%  # 15 días antes
                 bind_rows(tibble(date = rep(incendio_estudio$FECHA_INIC-30,length(grid)),geometry = grid)) %>% # el mes antes
  st_sf() 
```


```{r}
source("scripts/strat/fun_asignar_variables - copia.R")
full_grid <- asignar_variables(sample)

# Se elimininas las observaciones con valores faltantes:
full_grid <- full_grid %>% drop_na()
```

Se agrupan los niveles de uso de suelo
```{r}
full_grid <- full_grid |> 
  mutate(uso_suelo = fct_other(uso_suelo,
                               keep = c("21","22","23","24","31","32","33"),
                               other_level= "Otro"))

full_grid %>% st_drop_geometry() %>% skim()
# save(full_grid,file = "salidas_intermedias/full_grid_incendio_0921_processed.RData")
load("salidas_intermedias/full_grid_incendio_0921_processed.RData")
```

Se usará el modelo de regresión logística lasso final,
```{r}
# load("salidas_intermedias/trained_models_strat_2024_04_28.RData") # Modelos
load("Private/all_models_test.RData")

model <- models %>% filter(model_name=="lr")
# model <- models %>% filter(model_name=="svm_rbf")
# model <- models %>% filter(model_name=="svm_linear")
# model <- models %>% filter(model_name=="rf")

rm(models) # Es un archivo muy pesado, lo eliminamos de la memoria
```


```{r}
pred_class = model %>% 
  pull(last_fit) %>% 
  .[[1]] %>% 
  extract_workflow() %>% 
  predict(new_data = full_grid)
pred_probs = model %>% 
  pull(last_fit) %>% 
  .[[1]] %>% 
  extract_workflow() %>% 
  predict(new_data = full_grid,type="prob")
pred = cbind(full_grid,pred_class,pred_probs)
```

```{r}
g <- ggplot(data = and) + 
  geom_sf() +
  geom_sf(data = pred, aes(color=.pred_1, alpha = .pred_1),size = 1.5) +
  facet_wrap(~date) + 
  scale_color_gradientn(colours = rainbow(5,rev=T),limits=c(0,1)) +
  # scale_color_gradient(low="blue", high="red")+
  labs(title="Incendio de Sierra Bermeja",
       subtitle = paste0("Model: ",model$model_name)) + 
  guides(alpha = "none") + 
  theme_minimal() +
  theme(axis.text.x=element_blank(),
        axis.ticks.x=element_blank(),
        axis.text.y=element_blank(),
        axis.ticks.y=element_blank(),
        panel.grid.major = element_blank(),
        panel.grid.minor = element_blank()) +
  geom_sf(data = incendio_estudio, fill="transparent",col="red") +
  labs(color = "Probabilidad\nestimada de\nincendio") +
  coord_sf(xlim=c(bbox$xmin,bbox$xmax),ylim=c(bbox$ymin,bbox$ymax)) 
  

g 
```

```{r}
g <- ggplot(data = and) + 
  geom_sf() +
  geom_sf(data = pred, aes(color=.pred_class, alpha = .pred_1),size = 1.5) +
  facet_wrap(~date) + 
  # scale_color_gradientn(colours = rainbow(5,rev=T),limits=c(0,1)) +
  # scale_color_gradient(low="blue", high="red")+
  labs(title="Incendio de Sierra Bermeja",
       subtitle = paste0("Model: ",model$model_name)) + 
  guides(alpha = "none") + 
  theme_minimal() +
  theme(axis.text.x=element_blank(),
        axis.ticks.x=element_blank(),
        axis.text.y=element_blank(),
        axis.ticks.y=element_blank(),
        panel.grid.major = element_blank(),
        panel.grid.minor = element_blank()) +
  geom_sf(data = incendio_estudio, fill="transparent",col="red") +
  labs(color = "Probabilidad\nestimada de\nincendio") +
  coord_sf(xlim=c(bbox$xmin,bbox$xmax),ylim=c(bbox$ymin,bbox$ymax)) 
  

g 

```

```{r}
pred %>% 
  group_by(date) %>% 
  summarize(AVG.pred_1 = mean(.pred_1)) %>% 
  st_drop_geometry()
```


# --------------------------
```{r}
# Más detalle
grid = st_make_grid(st_buffer(incendio_estudio,dist=010000),
                    cellsize = c(1000,1000),
                    what = "centers")[and]

# ----------
g + 
  geom_sf(data = grid,size=0.7) +
  coord_sf(xlim=c(bbox$xmin,bbox$xmax),ylim=c(bbox$ymin,bbox$ymax))

# sample = st_sf(tibble(date = rep(incendio_estudio$FECHA_INIC,length(grid)),geometry = grid))

sample = tibble(date = rep(incendio_estudio$FECHA_INIC,length(grid)),geometry = grid) %>% # el día del incendio
                 bind_rows(tibble(date = rep(incendio_estudio$FECHA_INIC+15,length(grid)),geometry = grid)) %>%  # 15 días después
                 bind_rows(tibble(date = rep(incendio_estudio$FECHA_INIC+15,length(grid)),geometry = grid)) %>%  # el mes después
                 bind_rows(tibble(date = rep(incendio_estudio$FECHA_INIC-15,length(grid)),geometry = grid)) %>%  # 15 días antes
                 bind_rows(tibble(date = rep(incendio_estudio$FECHA_INIC-30,length(grid)),geometry = grid)) %>% # el mes antes
  st_sf() 


source("scripts/strat/fun_asignar_variables.R")
full_grid  = asignar_variables(sample)


# ----------

# Cargar lr_model en modelos
pred_class = last_lr_fit %>% extract_workflow() %>% predict(new_data = full_grid)
pred_probs = last_lr_fit %>% extract_workflow() %>% predict(new_data = full_grid,type="prob")
pred = cbind(full_grid,pred_class,pred_probs)

g + geom_sf(data = pred, mapping = aes(color=.pred_class))

bbox = st_bbox(st_buffer(incendio_estudio,dist=10000)) %>%  as.matrix() %>% t() %>% as.data.frame()

g + 
  geom_sf(data = pred, mapping = aes(color=.pred_class)) +
 #scale_color_gradientn(colours = rainbow(5,rev=T))  +
  geom_sf(data=incendio_estudio,fill = alpha("black",0)) +
  coord_sf(xlim=c(bbox$xmin,bbox$xmax),ylim=c(bbox$ymin,bbox$ymax))


g + 
  geom_sf(data = pred, mapping = aes(color=.pred_1)) +
  scale_color_gradientn(colours = rainbow(5,rev=T))  +
  geom_sf(data=incendio_estudio,fill = alpha("black",0)) +
  coord_sf(xlim=c(bbox$xmin,bbox$xmax),ylim=c(bbox$ymin,bbox$ymax))

# Lógico ya que es un problema muy complejo y la resolución de las variables meteorológicas es de alrededor de 50km, por lo que intentar hacer predicciones a una escala menor (en este caso 50 veces menor) no es una buena idea
# Además el incendio duró 46 días
```




# Otras cosas pa borrar


```{r}
# grid de puntos 10km x 10km de andalucía
grid = st_make_grid(and,
                    cellsize = c(25000,25000), # 10000
                    what = "centers")[and]

g + geom_sf(data = grid)
```


```{r}
sample = st_sf(tibble(date = rep(incendio_estudio$FECHA_INIC,length(grid)),geometry = grid) %>% # el día del incendio
                 bind_rows(tibble(date = rep(incendio_estudio$FECHA_INIC+15,length(grid)),geometry = grid)) %>%  # 15 días después
                 bind_rows(tibble(date = rep(incendio_estudio$FECHA_INIC+15,length(grid)),geometry = grid)) %>%  # el mes después
                 bind_rows(tibble(date = rep(incendio_estudio$FECHA_INIC-15,length(grid)),geometry = grid)) %>%  # 15 días antes
                 bind_rows(tibble(date = rep(incendio_estudio$FECHA_INIC-30,length(grid)),geometry = grid))) # el mes antes


source("scripts/strat/fun_asignar_variables.R")
full_grid  = asignar_variables(sample)

full_grid
```

Se usará el modelo de regresión lineal entrenado con los datos de entrenamiento y validación

```{r}
# Cargar lr_model en modelos
model
pred_class = last_lr_fit %>% extract_workflow() %>% predict(new_data = full_grid)
pred_probs = last_lr_fit %>% extract_workflow() %>% predict(new_data = full_grid,type="prob")
pred = cbind(full_grid,pred_class,pred_probs)

g + geom_sf(data = pred, mapping = aes(color=.pred_class))

g + 
  geom_sf(data = pred, mapping = aes(color=.pred_1)) +
  scale_color_gradientn(colours = rainbow(5,rev=T)) 
```
